+++ src/Controller/Adminhtml/System/Config/Payment/Refresh.php
@@ -64,23 +64,7 @@ class Refresh extends \Magento\Backend\App\Action
          */
         $resultJson = $this->resultJsonFactory->create();
 
-        $request = $this->getRequest();
-
-        if($request->getParam('is_live_mode')) {
-            $merchantAgremmentCode = $request->getParam('merchant_agreement_code', null);
-            $shopPrivateKeyfile = $request->getParam('shop_private_keyfile', null);
-            $payPagePublicKeyfile = $request->getParam('pay_page_public_keyfile', null);
-        } else {
-            $merchantAgremmentCode = $request->getParam('merchant_agreement_code_test', null);
-            $shopPrivateKeyfile = $request->getParam('shop_private_keyfile_test', null);
-            $payPagePublicKeyfile = $request->getParam('pay_page_public_keyfile_test', null);
-        }
-
-        $paymentMethods = $this->_paymentMethodHelper->refreshPaymentMethods(
-            $merchantAgremmentCode,
-            $shopPrivateKeyfile,
-            $payPagePublicKeyfile
-        );
+        $paymentMethods = $this->_paymentMethodHelper->refreshPaymentMethods();
 
         if (is_string($paymentMethods)) {
             return $resultJson->setData(