+++ src/Model/Client/Rest/Config.php
@@ -21,23 +21,22 @@ class Config extends \Verifone\Payment\Model\Client\Config
      * Config constructor.
      *
      * @param \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig
-     * @param \Magento\Store\Model\StoreManagerInterface         $storeManager
-     * @param \Magento\Framework\App\ProductMetadataInterface    $productMetadata
-     * @param \Magento\Framework\App\Filesystem\DirectoryList    $directoryList
-     * @param \Verifone\Payment\Helper\Payment                   $helper
-     * @param \Magento\Framework\UrlInterface                    $urlBuilder
+     * @param \Magento\Store\Model\StoreManagerInterface $storeManager
+     * @param \Magento\Framework\App\ProductMetadataInterface $productMetadata
+     * @param \Magento\Framework\Module\Dir\Reader $reader,
+     * @param \Verifone\Payment\Helper\Payment $helper
+     * @param \Magento\Framework\UrlInterface $urlBuilder
      */
     public function __construct(
         \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
         \Magento\Store\Model\StoreManagerInterface $storeManager,
         \Magento\Framework\App\ProductMetadataInterface $productMetadata,
-        \Magento\Framework\App\Filesystem\DirectoryList $directoryList,
+        \Magento\Framework\Module\Dir\Reader $reader,
         \Verifone\Payment\Helper\Payment $helper
-    ) {
+    )
+    {
 
-        parent::__construct($scopeConfig, $storeManager, $productMetadata, $directoryList, $helper);
-
-        return $this;
+        parent::__construct($scopeConfig, $storeManager, $productMetadata, $reader, $helper);
 
     }